fix and simplify storybook configuration significantly.
