+++ apps/vor/frontend/package.json
@@ -49,17 +49,9 @@
   "devDependencies": {
     "@babel/preset-typescript": "7.10.4",
     "@istanbuljs/nyc-config-typescript": "1.0.1",
-    "@storybook/addon-a11y": "6.0.21",
     "@storybook/addon-actions": "6.0.21",
-    "@storybook/addon-docs": "6.0.21",
-    "@storybook/addon-knobs": "6.0.21",
-    "@storybook/addon-links": "6.0.21",
-    "@storybook/addon-storysource": "6.0.21",
-    "@storybook/addon-viewport": "6.0.21",
-    "@storybook/addons": "6.0.21",
+    "@storybook/addon-essentials": "^6.0.21",
     "@storybook/react": "6.0.21",
-    "@storybook/storybook-deployer": "2.8.6",
-    "@storybook/theming": "6.0.21",
     "@testing-library/react": "10.4.9",
     "@testing-library/react-hooks": "3.4.1",
     "@types/date-fns": "2.6.0",
@@ -99,7 +91,6 @@
     "jest-date-mock": "1.0.8",
     "jest-emotion": "10.0.32",
     "jest-fetch-mock": "3.0.3",
-    "miragejs": "0.1.40",
     "npm-run-all": "4.1.5",
     "prettier": "2.1.1",
     "react-docgen-typescript-loader": "3.7.2",
