+++ apps/vor/frontend/src/components/LinkInput/LinkInput.stories.tsx
@@ -1,5 +1,4 @@
 import React, { FC } from "react"
-import { boolean } from "@storybook/addon-knobs"
 import { action } from "@storybook/addon-actions"
 import LinkInput from "./LinkInput"
 
@@ -13,7 +12,7 @@ export default {
 
 export const Basic: FC = () => (
   <LinkInput
-    isLoading={boolean("loading", false)}
+    isLoading={false}
     onChange={action("onChange")}
     onSave={action("onSave")}
   />
